requirements.txt
aiohttp==3.6.2
appdirs==1.4.3
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
CacheControl==0.12.6
cachetools==4.1.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
colorlog==4.2.1
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.0
distro==1.4.0
Django==3.1.1
django-cors-headers==3.5.0
django-json-widget==1.0.1
djangorestframework==3.11.1
drf-yasg==1.17.1
future==0.18.2
geographiclib==1.50
geopy==2.0.0
google-api-core==1.22.2
google-auth==1.22.0
google-cloud-pubsub==2.1.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.32.0
gunicorn==20.0.4
html5lib==1.0.1
humanfriendly==8.2
idna==2.8
inflection==0.5.1
ipaddr==2.2.0
itypes==1.2.0
Jinja2==2.11.2
libcst==0.3.11
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
multidict==4.7.6
mypy-extensions==0.4.3
openapi-codec==1.3.2
packaging==20.3
pep517==0.8.2
polyline==1.4.0
progress==1.5
proto-plus==1.10.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
python-decouple==3.3
pytoml==0.1.21
pytz==2020.1
PyYAML==5.3.1
requests==2.22.0
retrying==1.3.3
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.14.0
sqlparse==0.3.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1
yarl==1.6.0
